requirements.txt
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.10
python-decouple==3.0
static3==0.7.0
gunicorn==19.4.1
psycopg2==2.6.1
django_extensions==1.6.1
django_test_without_migrations==0.4
pytz
pdfminer3k
PyDrive
django_tables2
PyPDF2